pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "mdscope"
version = "0.1.0"
description = "User-friendly MD trajectory analysis automation powered by MDAnalysis"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "mdscope contributors" }]
dependencies = [
  "typer>=0.12",
  "pydantic>=2.5",
  "pyyaml>=6.0",
  "MDAnalysis>=2.6,<3.0",
  "numpy>=1.24",
  "scipy>=1.10",
  "pandas>=2.0",
  "matplotlib>=3.7",
  "scikit-learn>=1.3",
  "biopython>=1.83",
  "hdbscan>=0.8.33",
]

[project.optional-dependencies]
analysis = ["tqdm>=4.66"]
sasa = ["mdakit-sasa>=0.2.0"]
dev = ["pytest>=8.0", "ruff>=0.6", "build>=1.2"]

[project.scripts]
mdscope = "mdscope.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/mdscope"]
